+++ src/constants/changelog.ts
@@ -2,6 +2,29 @@ import { TReleaseNotes } from 'constants/types';
 
 const CHANGELOG: TReleaseNotes = [
   {
+    name: 'Version 3.15.0',
+    date: 1637597204585,
+    notes: [
+      {
+        description:
+          'App is now available in german and spanish, can be changed from the options page',
+        type: 'FEATURE',
+      },
+      {
+        description: 'App is now available for iOS in the App store!',
+        type: 'UPDATE',
+      },
+      {
+        description: "Added Pikachu to starter's list for Red, Blue and Yellow",
+        type: 'FIX',
+      },
+      {
+        description: "Changed Norman's second pokemon from Slaking to Vigoroth in Emerald",
+        type: 'FIX',
+      },
+    ],
+  },
+  {
     name: 'Version 3.14.2',
     date: 1637597204585,
     notes: [